+++ common/src/main/java/com/telepathicgrunt/the_bumblezone/modcompat/ModChecker.java
@@ -18,6 +18,7 @@ public class ModChecker {
 	public static final List<ModCompat> BLOCK_TELEPORT_COMPATS = new ArrayList<>();
 	public static final List<ModCompat> COMB_ORE_COMPATS = new ArrayList<>();
 	public static final List<ModCompat> HOST_BEE_COMPATS = new ArrayList<>();
+	public static final List<ModCompat> PROJECTILE_IMPACT_HANDLED_COMPATS = new ArrayList<>();
 
 	public static boolean productiveBeesPresent = false;
 	public static boolean resourcefulBeesPresent = false;
@@ -29,6 +30,7 @@ public class ModChecker {
 	public static boolean potionOfBeesPresent = false;
 	public static boolean requiemPresent = false;
 	public static boolean twilightForestPresent = false;
+	public static boolean dragonEnchantPresent = false;
 
 	/*
 	 * -- DO NOT TURN THE LAMBDAS INTO METHOD REFS. Method refs are not classloading safe. --
@@ -71,6 +73,7 @@ public class ModChecker {
 				if (compat.compatTypes().contains(ModCompat.Type.BLOCK_TELEPORT)) BLOCK_TELEPORT_COMPATS.add(compat);
 				if (compat.compatTypes().contains(ModCompat.Type.COMB_ORE)) COMB_ORE_COMPATS.add(compat);
 				if (compat.compatTypes().contains(ModCompat.Type.HAS_HOST_BEES)) HOST_BEE_COMPATS.add(compat);
+				if (compat.compatTypes().contains(ModCompat.Type.PROJECTILE_IMPACT_HANDLED)) PROJECTILE_IMPACT_HANDLED_COMPATS.add(compat);
 			}
 		}
 		catch (Throwable e) {