+++ forge/src/main/java/com/telepathicgrunt/the_bumblezone/modcompat/forge/ForgeModChecker.java
@@ -41,6 +41,9 @@ public class ForgeModChecker {
 
             modid = "potionofbees";
             loadupModCompat(modid, () -> new PotionOfBeesCompat());
+
+            modid = "dragon_enchants";
+            loadupModCompat(modid, () -> new DragonEnchantCompat());
         }
         catch (Throwable e) {
             printErrorToLogs("classloading " + modid + " and so, mod compat done afterwards broke");